+++ entrypoint
@@ -25,4 +25,4 @@ echo " "
 echo "==================================================================================================="
 echo "==================================================================================================="
 
-python /opt/conda-forge/rerender.py
+run-webservices-dispatch-action
