+++ webservices_dispatch_action/tests/test_exists.py
@@ -1,0 +1,7 @@
+"""
+test to make sure things are running
+"""
+
+
+def test_exists():
+    assert True
